database_manager/requirements.txt
flask
mysql.connector
requests
